ml_service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
torch==2.0.1
pydantic==1.8.2
kafka-python==2.0.2
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
prometheus-client==0.11.0